add icon with SettingGame create TimeViewModel
